fix(persona): moved to homepage from courses page
